Add files via upload
Store sales and profit analysis is the task of analyzing the performance of a retail store in terms of its sales and profits. It helps businesses identify areas for improvement and make data-driven decisions to optimize their operations, pricing, marketing, and inventory management strategies to drive revenue and growth. So, if you want to learn how to analyze the sales and profit of a store, this article is for you. In this article, I will take you through the task of Store sales and profit analysis using Python.